backend/requirements.txt
# Core Flask dependencies
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
gunicorn==23.0.0
blinker==1.9.0
PyJWT==2.10.1

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.11  # PostgreSQL adapter for Python
alembic==1.16.5
greenlet==3.2.4
Mako==1.3.10

# HTTP requests and API calls
requests==2.32.4
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0

# QR Code generation
qrcode==7.4.2
Pillow==11.2.1

# Environment configuration
python-dotenv==1.0.1

# Additional utilities
Werkzeug==3.1.3
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
colorama==0.4.6
packaging==25.0
typing_extensions==4.15.0

# Test dependencies
pytest==8.4.2
iniconfig==2.3.0
pluggy==1.6.0
Pygments==2.19.2